Create Larger Summon Ring On Condition
If the behemoth vole is brought down to 30% or below its max hp before
it spawns the first ring of voles, then a secondary larger ring of
Giant Voles will spawn around the player in addition to the smaller
first ring. Also, increased the damage behemoth can do by 5 points.