Add New Projectile Attack to Behemoth Vole Moveset
Added a new projectile attack for the behemoth vole to do. However, need
to include use of the projectile container node in the vomit_projectile
scene. Also, need to make projectile cause a poisoned status effect on
hit players, and also create a puddle of poison at the where the
projectile dies.