Create Poison Puddle At End of Projectile Lifetime
Once the vomit projectile that the boss shoots has reached the end of its lifetime,
a new sprite will replace the projectile one indicating the it is now a vomit puddle.
This puddle is poisonous. Also, I changed the visibility layer of all enemies to
be top level, along with the initial poison projectile. This way, once the projectile turns
into a puddle, I can change the visibility layer of the puddle to be behind every enemy.