Add New Split Body Move to Transient Wraith Boss
The transient wraith is now capable of splitting into 3 other clones,
each with half the current HP of the wraith spawning the clone. This
feature is unfinished and likely very unbalanced. Instead of making
this a special attack that may occur every couple seconds, I may just
include this move as a part of the boss's second phase.