Make Dungeon Child of Meadow
Instead of making the meadow dungeon a scene to be instantiated in script,
it is now implemented as a child of the meadow node. Also, changed some hard
coded node locations to be more variable; the places where the player spawns
into and out of the dungeon is now attached to the location of the dungeon
child nodes. If the dungeon is ever placed elsewhere in the global plane,
no code should need to be updated to reflect this (the transient wraith's
teleport restrictions still need to be made variable, however). Also, I added
a little decoration outside of the dungeon, and intend to add a little more.