Store Wraith's Spawn Location & Arena Measurements
Started working on making the boundaries for the wraith's teleportation
locations variable. I intend to export the x length and y width of
the boss's arena as variables that can be modified whenever.
This way, all that needs to be done any time the arena measurements change
is to just quickly measure the entire length & width of arena (arena must
be imagined as a rectangle whether it is or isn't) and change the respective
variables. In the boss script, I'll use these variables to figure out the max
distance the wraith can teleport to so it doesn't leave the arena. As it is
now, these values are hard coded in, making it annoying to have to change
every time the arena measurement change.