Make Wraith Teleport Boundary Based on Arena Size
Instead of the transient wraith's teleportation boundaries hard-coded
values, it's now calculated based on the x and y lengths of the arena
it's in. These are stored in exported variables so they may be easily
changed in the future if need be. Also included a variable that will
factor in an extra amount of space to leave between maximum teleport
spot and wall