Change Spawn Condition & Rebalance for Vole Boss
The behemoth vole now spawns under different conditions and in a different
location. Also, the boss has been heavily reworked to provide more of a challenge,
and to scale off the player level instead of how much damage is dealt to it.
The transient wraith boss also now has double its maximum health.